Change contents of %local-nicknames to include a package object
Previously, %local-nicknames basically contained a list of two
strings.  Now, the actual package is stored as a package object.  This
requires updating all the code to handle that.  And for
`package-local-nicknames`, we return a copy of the value of
`%local-nicknames`.  Not sure if that's necessary, but we're then sure
the nicknames can't be accidentally modified.

Also updated `delete-package` to remove any local nicknames from any
package.  Not really tested.