Add some new variables so we have the correct values everywhere
Add variables "year" and "month" so we have the correct year and month
of the snapshot we want. (We'll have to do something different when we
use a release.)

Use these variables in the download_url and the version.

Finally, add variable "tar_ext" so we use the correct tar extension
everywhere too.