Replace __HI and __LO macros that no longer exist
The __HI and __LO macros to access the high and parts of a double no
longer exist.  We need to use a union for this.

To make the code look like the original, define some new inline
functions get_hiword, get_loword, and set_hiword.  The first two get,
respectively, the high and low words of a double float.  The second
sets the high word of a double to the given value.  These are all
that's needed for the hypot function.