Update lisp to use the new correctly rounded functions.
Update GNUmakefile to compile the special functions we need.  These
are added into the lisp executable (and lisp.a).

In Config.x86_linux enable support for errno in core-math routines so
that errno is set appropriately.  We enable this for compatibility
with how the routines used to work.

Update irrat.lisp to use the new correctly rounded functions.