The float-modes type is a 30 bit unsigned byte
Since we use a single 32-bit word to set the x87 FP modes, we need at
least 30 bits because the top 2 bits are reserved so they shouldn't be
set.  This is because the control word is in the top 16 bits.
Previously, the float-modes type was an (unsigned-byte 24), which is
too small now.