Fix #426:  Define float-modes type correctly
The definition of `float-modes` is `(unsigned-byte 24)`.  This is too
small if we are trying to set the precision control field for x87 via
`(setf (x86::x87-floating-point-modes) ...)`.

The field needs to be at least 30 bits to that we can assign values
for the x87 control and status words.  Thus, set the type to
`(unsigned-byte 32)`.