Set the x87 precision to double when calling out to C
When we call out to C, we save the x87 FPU control word.  Then update
the FPU control to have the precision set to double (53-bit)
precision.  This is needed for the core-math routines that depend on
the x87 precision control to be double.

When the function returns, we restore the original x87 control word.