Define new temp to old the FPU Cw
We were using edx as a temp to old the FPU CW, but that might
interfere with the actual use of the edx register.  This showed up
with a segfault in computing (abs #c(-3d0 4d0)) (from the ansi-tests).

Hence, define a new temp reg that we can for the FPU CW.