Register :core-math as a runtime feature for Linux
Add a new feature :core-math to enable the use of the core-math
functions.  This is so we can continue to run CI on Darwin until we
can figure out if we can use core-math on the ancient version of
Darwin that we're using for CI.

Thus, set the :core-math feature in Linux-os.lisp.  Update irrat.lisp
to use core-math only if :core-math is defined.  Otherwise, we use
fdlibm, as before.

Update Config.x86_common to define the core-math object files when
FEATURE_CORE_MATH is defined (because we registered :core-math as a
runtime feature).  Move the core-math object files from GNUmakefile to
Config.x86_linux_common.