Modify code to call the single-float functions as needed.
The `def-math-rtn` macro now defines interfaces for the double and
single precision core-math routines.  Updated handle-reals to call the
single or double precision routines.

Manually tested this with sinf by using gdb to set a breakpoint at
cr_sinf and running `(sin 1f0)` in lisp.  We hit the breakpoint as
expected.