Fix #355:  Update sigfpe-handler for solaris/x86
The `sigfpe-handler` didn't recognize SIGFPE signals on solaris/x86.
On solaris/x86, the `code` arg to `sigfpe-handler` contains the info
to determine what caused the SIGFPE.

With this change, things like `(exp 1000d0)` correctly signal an
overflow and when we use the debugger to return to the REPL,
`(get-floating-point-modes)` shows that we have previous settings were
preserved as expected.