Revert previous change and use Madhu's solution
Madhu provided this patch to fix the issue he was encountering with
probe-file that was calling unix-stat which was failing for him.  (I
don't have that problem for whatever reason.)

This changes `time-t` to be an `int64-t` type and adds
`-D_TIME_BITS=64` to the compilation flags on Linux.

I initially had issues with this patch, but it was fixed by removing
all the build directories.