Fix #444: Use 64-bit time_t value for stat and friends
Define `_TIME_BITS=64` in os-common.c so that `stat` and friends will
use a 64-bit `time_t` type.  This is currently only done for Linux.

Manually tested this on a test file and we return the correct mtime
and ctime values.

See #444 for the test file and C test program to verify the mtime and
ctime values.