Address review comments
Add test for timestamp in 2106 that won't fit in a 32-bit int.  To do
this, we added a new file 64-bit-timestamp-2106.txt and renamed
64-bit-timestamp.txt to 64-bit-timestamp-2038.txt.

Update run-unit-tests.sh to set the time stamps on these two
files, and add a test to os.lisp for the new file.

Finally, a minor tweak to os.lisp because we got the expected and
actual values backwards in `assert-equal` for the time stamp tests.
The expected value is first.