Put the errno symbols in the UNIX package
Instead of having a separate ERRNO package holding all the ERRNO
symbols, we're rearranging this and putting them all in the UNIX
package as they used to be.

Thus, add a new file unix-pkg-template.lisp to hold all entire
definition of the UNIX package except for the errno symbols.  Update
create-errno.sh to create the new file exports-unix.lisp from this
template.

errno-template.lisp is updated so that the symbols are placed in the
UNIX package instead of the ERRNO package.

Update bin/errno-linux.lisp with the new sorted list, which is in
reverse so that the aliases come at the end.

Update exports.lisp to remove the UNIX package and ERRNO package
definitions.

Update worldcom and worldload to load/compile exports-unix.lisp at the
after loading/compiling exports.lisp.