Fix #453:  Use correct flags when running the analyzer
When running the analyzer, we previously set `CFLAGS=-fanalyzer`.
However, this overrides whatever settings we had in the config files
and makefile for CFLAGS.  This means we no longer set
`_FILE_OFFSET_BITS=64`.  Hence, we got an error in the analysis.

Add a new variable `ANALYZER` that is set to `-fanalyzer` when doing
the analyzer step of CI.