Clean up implementation and fix contagion issues
First, just define `cdiv-double-float` that implements Baudin and
Smith for (complex double-float).  For (complex single-float), convert
to double-float and use Smith's algorithm instead of the more complex
Baudin and Smith.

Fix up the numeric contagion so we call these two routines
appropriately.  (This is quite messy!)

Update the deftransforms to call kernel::cdiv-double-float and
kernel::cdiv-single-float.