Fix bug in cdiv-single-float and add/update tests
cdiv-single-float needs to convert single-float to double-float so we
don't overflow.

Update the accuracy of the one test to be 52 instead of 53.  The
equivalent C code has the same accuracy.

Add a test that we don't overflow for division for complex
single-float number.

Finally, add tests to make sure we got the all the numeric contagion
cases covered for division.  Quite useful for getting the contagion
cases covered.