Add declarations and set speed 3 and safety 0
Set speed 3 and safety 0 for `cdiv-double-float`.  This causes lots of
compiler warnings, so fix those.  The means adding some declarations
because the compiler can't figure out the type of the numbers.

Also change `maybe-scale` to include the components of the complex
numbers.  The possibly updated values are returned as multiple values.
Update callers appropriately.  This gets rid of some extraneous boxing
notes.