Fix #458:  Fix spurious overflow in double-double multiply
We were overflowing when computing `(* -2w300 1w0)` instead of
returning `-2w300`.  But `(* 2w300 1w0)` returned the correct value.

This is a bug in `c::two-prod` where we were checking if a value was
greater than `+two970+` ($2^{-970}$).  This of course fails if the
value is negative.  Hence we need to take the absolute value before
testing.

Add a simple test that we return the correct value for `(* -2w300
1w0)`.